public/requirements.txt
Flask
flask-cors
smbus2